draw_key {ggstance}	R Documentation

Horizontal key drawing functions

Description

Horizontal key drawing functions

Usage

draw_key_hpath(data, params, size)

draw_key_pointrangeh(data, params, size)

draw_key_crossbarh(data, params, size)

draw_key_boxploth(data, params, size)

Arguments

data
A single row data frame containing the scaled aesthetics to display in this key

params
A list of additional parameters supplied to the geom.

size
Width and height of key in mm.

Value

A grid grob.

[Package ggstance version 0.3.7 Index]