geom_crossbarh {ggstance}	R Documentation

Horizontal intervals: lines, crossbars & errorbars.

Description

Horizontal versions of geom_linerange(), geom_pointrange(), geom_errorbar() and geom_crossbar().

Usage

geom_crossbarh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 2.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_errorbarh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_linerangeh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_pointrangeh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 4,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat
The statistical transformation to use on the data for this layer, as a string.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

fatten
A multiplicative factor used to increase the size of the middle bar in geom_crossbar() and the middle point in geom_pointrange().

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Aesthetics

geom_crossbarh() understands the following aesthetics (required aesthetics are in bold):

x

y

xmin

xmax

alpha

colour

fill

group

linetype

size

Learn more about setting these aesthetics in vignette("ggplot2-specs").

geom_errorbarh() understands the following aesthetics (required aesthetics are in bold):

y

xmin

xmax

alpha

colour

group

linetype

size

width

Learn more about setting these aesthetics in vignette("ggplot2-specs").

geom_linerangeh() understands the following aesthetics (required aesthetics are in bold):

y

xmin

xmax

alpha

colour

group

linetype

size

Learn more about setting these aesthetics in vignette("ggplot2-specs").

geom_pointrangeh() understands the following aesthetics (required aesthetics are in bold):

x

y

xmin

xmax

alpha

colour

fill

group

linetype

shape

size

stroke

Learn more about setting these aesthetics in vignette("ggplot2-specs").

Different between ggplot2 and ggstance

'ggplot2::geom_errorbarh()' uses the 'height' aesthetic. The ggstance version uses the 'width' aesthetic. This is for consistency with the direction of the geom and other ggstance functions. You can still supply 'height' for compatibility.

[Package ggstance version 0.3.7 Index]