hmisc_h {ggstance}	R Documentation

Horizontal versions of summary functions from Hmisc

Description

These are horizontal versions of the wrappers around functions from Hmisc designed to make them easier to use with stat_summaryh. The corresponding vertical versions are hmisc(). See the Hmisc documentation for more details:

smean.cl.boot

smean.cl.normal

smean.sdl

smedian.hilow

Usage

mean_cl_boot_h(x, ...)

mean_cl_normal_h(x, ...)

mean_sdl_h(x, ...)

median_hilow_h(x, ...)

Arguments

x
a numeric vector

...
other arguments passed on to the respective Hmisc function.

Value

A data frame with columns x, xmin, and xmax.

Examples

if (requireNamespace("Hmisc")) {
  x <- rnorm(100)
  mean_cl_boot_h(x)
  mean_cl_normal_h(x)
  mean_sdl_h(x)
  median_hilow_h(x)
}

[Package ggstance version 0.3.7 Index]